Establish Arduino mock environment, Python simulation, and performance analysis.
This commit introduces:
- A C++ mock Arduino environment in 'mock_arduino/'.
- A Python simulation script 'simulator.py' for signal and frequency analysis.
- Transfer function performance graphs in 'transfer_function.png'.
- Detailed README updates explaining implementations and simulation.
- Critical bug fixes in .ino files: 16-bit overflows, division-by-zero protection, and corrected Timer1 PWM modes.

Intermediate compilation files are generated in /tmp to maintain a clean repository.

+++ class_D_passthrough.ino
@@ -21,7 +21,7 @@
 #define MIN_PERIOD 976 // microseconds
 
 // Define a variable to store the current sampling period
-unsigned int period;
+unsigned long period;
 
 // Define a variable to store the timer overflow count
 unsigned long timer;
